refactor(tab-set): rename rkTabView component to rkTabSet & docs update
