+++ example/App.js
@@ -26,10 +26,10 @@ const ExplorerApp = StackNavigator({
   HorizontalCalendar: { screen: Screens.HorizontalCalendarScreen },
   Switch: { screen: Screens.SwitchScreen },
   Choice: { screen: Screens.ChoiceScreen },
-  TabView: { screen: Screens.TabsScreen },
-  BaseTabs: { screen: Screens.BaseTabViewScreen },
-  BadgeTabs: { screen: Screens.BadgeTabViewScreen },
-  IconTabs: { screen: Screens.IconTabsScreen },
+  TabSet: { screen: Screens.TabsScreen },
+  BaseTabs: { screen: Screens.BaseTabScreen },
+  BadgeTabs: { screen: Screens.BadgeTabScreen },
+  IconTabs: { screen: Screens.IconTabScreen },
   Card: { screen: Screens.CardScreen },
   Avatar: { screen: Screens.AvatarScreen },
   Input: { screen: Screens.InputScreen },
