+++ example/screens/index.js
@@ -12,10 +12,10 @@ export * from './ComponentsScreen';
 export * from './ImageScreen';
 export * from './GalleryScreen';
 export * from './InputScreen';
-export * from './tabview/TabScreen';
-export * from './tabview/BaseTabViewScreen';
-export * from './tabview/BadgeTabViewScreen';
-export * from './tabview/IconTabsScreen';
+export * from './tabset/TabsScreen';
+export * from './tabset/BaseTabScreen';
+export * from './tabset/BadgeTabScreen';
+export * from './tabset/IconTabScreen';
 export * from './SettingsScreen';
 export * from './ChoiceCustomizationScreen';
 export * from './PickerScreen';
