+++ index.js
@@ -8,7 +8,8 @@ export { RkModalImg } from './src/components/image/rkModalImg';
 export { RkGallery } from './src/components/gallery/rkGallery';
 export { RkGalleryImage } from './src/components/gallery/rkGalleryImage';
 export { RkCalendar } from './src/components/calendar/rkCalendar.component';
-export { RkTabView } from './src/components/tabset/rkTabView.component';
+export { RkTabView } from './src/components/tab/rkTabView';
+export { RkTabSet } from './src/components/tabset/rkTabSet.component';
 export { RkTab } from './src/components/tabset/rkTab.component';
 export { RkCard } from './src/components/card/rkCard';
 export { RkAvoidKeyboard } from './src/components/avoidKeyboard/rkAvoidKeyboard';
