+++ src/components/tabset/rkTabBar.component.js
@@ -9,7 +9,17 @@ import {
 import { RkComponent } from '../rkComponent';
 
 /**
+ * `RkTabBar` is a component that manages `RkTab`s.
+ *
  * @extends React.Component
+ *
+ * @property {React.ReactNode} children - `RkTab` components with it's contents.
+ * @property {number} selectedIndex - Selected RkTab index.
+ * @property {bool} isScrollable - Defines if container should be scrollable.
+ * For cases with lots of tabs.
+ * @property {function} onSelect - Fires when `RkTab` is selected.
+ * @property {number} componentWidth - width of `RkTabBar`.
+ * Needed for `RkTab` equal distribution.
  */
 export class RkTabBar extends RkComponent {
   static propTypes = {
@@ -45,24 +55,30 @@ export class RkTabBar extends RkComponent {
   };
 
   /**
-   * @param params - object: { offset: number, animated: boolean }
+   * scrolls container to passed index
+   *
+   * @param params - object: { index: number, animated: boolean }
    */
-  scrollToIndex = (params) => {
+  scrollToIndex(params) {
     const offset = (this.props.componentWidth / this.props.children.length) * params.index;
     this.scrollToOffset({ offset });
   };
 
   /**
+   * scrolls container to passed offset
+   *
    * @param params - object: { offset: number, animated: boolean }
    */
-  scrollToOffset = (params) => {
+  scrollToOffset(params) {
     this.containerRef.scrollTo({ x: params.offset, ...params });
   };
 
   /**
+   * scrolls container to end
+   *
    * @param params - object: { animated: boolean }
    */
-  scrollToEnd = (params) => {
+  scrollToEnd(params) {
     this.containerRef.scrollToEnd(params);
   };
